api/HTTP/Controllers/Auth/oauth2.py
from datetime import datetime, timedelta
from typing import Union

from fastapi import HTTPException, status, Depends
from fastapi.security import OAuth2PasswordBearer
from jose import JWTError, jwt
from passlib.context import CryptContext

from api.Models.Token import TokenData
from api.Models.User import User
from helper.DB import check_user_exist
from helper.ENV import APP_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES


class Oauth2(object):
    pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
    oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")

    @staticmethod
    def login(username: str, password: str):
        user = Oauth2.authenticate_user(username, password)
        if not user:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Incorrect username or password",
                headers={"WWW-Authenticate": "Bearer"},
            )
        access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
        access_token = Oauth2.create_access_token(
            data={"sub": user.username}, expires_delta=access_token_expires
        )
        return {"access_token": access_token, "token_type": "bearer"}

    @staticmethod
    def create_access_token(data: dict, expires_delta: Union[timedelta, None] = None):
        to_encode = data.copy()
        if expires_delta:
            expire = datetime.utcnow() + expires_delta
        else:
            expire = datetime.utcnow() + timedelta(minutes=15)
        to_encode.update({"exp": expire})
        encoded_jwt = jwt.encode(to_encode, APP_KEY, algorithm=ALGORITHM)
        return encoded_jwt

    @staticmethod
    def authenticate_user(username: str, password: str) -> bool or dict:
        user = check_user_exist(username, get_user_if_exist=True)
        if not user:
            return False
        user = User(**user)
        if not Oauth2.verify_password(password, user.password):
            return False
        return user

    @staticmethod
    def verify_password(plain_password, hashed_password):
        return Oauth2.pwd_context.verify(plain_password, hashed_password)

    @staticmethod
    async def get_current_user(token: str = Depends(oauth2_scheme)):
        credentials_exception = HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Could not validate credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )
        if not token:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="You must login.",
                headers={"WWW-Authenticate": "Bearer"},
            )
        try:
            payload = jwt.decode(token, APP_KEY, algorithms=[ALGORITHM])
            username: str = payload.get("sub")
            if username is None:
                raise credentials_exception
            token_data = TokenData(username=username)
        except JWTError:
            raise credentials_exception
        user = check_user_exist(username=token_data.username, get_user_if_exist=True)
        if user is None:
            raise credentials_exception
        return token_data
